Deprecate Clustered(DOM)DataTreeChangeListener
The semantics of ClusteredDOMDataTreeChangeListener is the default we
want to expose from DataTreeChangeExtension. Any integration with
clusters, or any other details, really, should happen through a
dedicated DOMDataBroker.Extension.

Redefine DataTreeChangeExtension, retaining compatibility methods with
migration guidance.

This cascades to ClusteredDataTreeChangeListener, which is just a
mdsal-binding-api mirror of the DOM concept.

JIRA: MDSAL-819
Change-Id: Ice4e1db64d34241115226d532b226cbdff8e045c
Signed-off-by: Ruslan Kashapov <ruslan.kashapov@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>