Fix component/service shutdown order
AdaptingTracker.removedService() is invoked when are losing a service
reference. The order of operations is significant here: we need to
deactivate the component before we unget the service.

The reason is that all of component's users need to be deactivated before
the service is allowed to be deactivate -- which may happen inside our call
to ungetService().

JIRA: MDSAL-877
Change-Id: If4361d350f8237ca8a19ef165c36ddca9109dfe2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 1dceb47953a4da73705586278621b5713f45b4b7)