WIP: Refactor DataObjectModification
DataObjectModification requires a rather funky dispatch in downstream
users: they need to check modificationType() and then have to make
assumptions as to what the nullability of dataBefore()/dataAfter() is.

Since we have Java 21, we can model the domain using a sealed interface
hierarchy and properly expose the nullability concerns.

// FIXME: final implementation bits
// FIXME: reconcile nullability with DataTreeCandidateNode, as CDS does
//        not provide before-image at all!

JIRA: MDSAL-889
Change-Id: I568a7db7e0405b56c8ea0c5e1d0c3183088ade00
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>