Add DataObjectModification.coerceKeyStep()
Our downstream users are routinely using dataBefore() to acquire an
EntryObject's key from DataObjectModification. Alternatively, they have
home-grown utility methods to check DataObjectModification.step() to
ensure it is a KeyStep.

This patch provides a simple centralized alternative, so it is at users'
fingertips as it is safer and more performant than dataBefore().

JIRA: MDSAL-892
Change-Id: I6fad9ccc98e29ee82ab90110a99b14151dcbb58a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>